mergeV/E
